added status command; redefined the related function in the command class
