+++ .coding_agent_memory.json
@@ -1,5 +1,5 @@
 {
   "conversation_history": [],
   "project_context": {},
-  "last_updated": "2025-10-03T20:21:41.695423"
+  "last_updated": "2025-10-03T20:56:38.457099"
 }
