--- backups/generated_function.py_20251003_155127.bak
@@ -1,9 +1,0 @@
-def generated_function():
-    """
-    A generated function based on your request.
-    
-    Returns:
-        None
-    """
-    # TODO: Implement functionality
-    pass
